refactor ExpectationSuiteViewer tests back to use fixtures
